remove  duckdb metric store old driver
