+++ weiser/drivers/metric_stores/__init__.py
@@ -1,11 +1,10 @@
 from typing import Union
 from weiser.drivers.metric_stores.duckdb import DuckDBMetricStore
-from weiser.drivers.metric_stores.duckdb_v2 import DuckDBMetricStoreV2
 from weiser.drivers.metric_stores.postgres import PostgresMetricStore
 from weiser.loader.models import MetricStore, MetricStoreType
 
 QUERY_TYPE_MAP = {
-    MetricStoreType.duckdb: DuckDBMetricStoreV2,
+    MetricStoreType.duckdb: DuckDBMetricStore,
     MetricStoreType.postgresql: PostgresMetricStore,
 }
 
@@ -15,10 +14,10 @@ MetricStoreDB = Union[DuckDBMetricStore, PostgresMetricStore]
 
 class MetricStoreFactory:
     @staticmethod
-    def create_driver(metric_store: MetricStore) -> MetricStoreDB:
+    def create_driver(metric_store: MetricStore, verbose: bool = False) -> MetricStoreDB:
         if metric_store:
             return QUERY_TYPE_MAP.get(
                 metric_store.db_type,
                 DuckDBMetricStore,
-            )(metric_store)
-        return DuckDBMetricStore(MetricStore())
+            )(metric_store, verbose=verbose)
+        return DuckDBMetricStore(MetricStore(verbose=verbose))
